New folder/requirements.txt
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-filter==2.1.0
djangorestframework==3.9.3
drf-yasg==1.15.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
numpy==1.16.3
pandas==0.24.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
ruamel.yaml==0.15.94
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.24.3
xlrd==1.2.0
